chore: reorganize unit tests for oauth package
