+++ packages/oauth/test/unit/commands/authorizations/rotate.unit.test.ts
