+++ packages/oauth/test/unit/commands/clients/rotate.unit.test.ts
